+++ src/routes/(main)/dashboard/settings/+page.svelte
@@ -8,7 +8,7 @@
   import type { SubmitFunction } from "@sveltejs/kit";
   import { cn } from "$lib/utils";
   import Separator from "$lib/components/ui/separator/separator.svelte";
-  import DashboardShell from "../(components)/dashboard-shell.svelte";
+  import LayoutShell from "../(components)/layout-shell.svelte";
 
   export let data: PageData;
 
@@ -28,7 +28,7 @@
   };
 </script>
 
-<DashboardShell heading="Inställningar" text="Hantera dina kontoinställningar.">
+<LayoutShell heading="Inställningar" text="Hantera dina kontoinställningar.">
   <form
     slot="content"
     method="post"
@@ -64,4 +64,4 @@
       </div>
     </div>
   </form>
-</DashboardShell>
+</LayoutShell>
